run_phantomjs {webdriver}	R Documentation

Start up phantomjs on localhost, and a random port

Description

Throws and error if phantom cannot be found, or cannot be started. It works with a timeout of five seconds.

Usage

run_phantomjs(debugLevel = c("INFO", "ERROR", "WARN", "DEBUG"),
  timeout = 5000)

Arguments

debugLevel
Phantom.js debug level, possible values: "INFO", "ERROR", "WARN", "DEBUG".

timeout
How long to wait (in milliseconds) for the webdriver connection to be established to the phantomjs process.

Value

A list of process, the callr::process object, and port, the local port where phantom is running.

[Package webdriver version 1.0.6 Index]