key {webdriver}	R Documentation

Special keys, so that we can refer to them with an easier syntax

Description

Special keys, so that we can refer to them with an easier syntax

Usage

key

Format

An object of class list of length 51.

Examples

## Not run: 
el$sendKeys("xyz")
el$sendKeys("x", "y", "z")
el$sendKeys("username", key$enter, "password", key$enter)

## Sending CTRL+A
el$sendKeys(key$control, "a")

## Note that modifier keys (control, alt, shift, meta) are sticky,
## they remain in effect in the rest of the sendKeys() call. E.g.
## this sends CTRL+X and CTRL+S
el$sendKeys(key$control, "x", "s")

## You will need multiple calls to release control and send CTRL+X S
el$sendKeys(key$control, "x")
el$sendKeys("s")

## End(Not run)

[Package webdriver version 1.0.6 Index]